// src/types/theme.interface.ts
export interface ITheme {
  // COLORS
  bgGradient: string;
  lightGray: string;
  gray: string;
  gray2: string;
  darkGray: string;
  graphite: string;
  black: string;
  blue: string;
  orange: string;
  red: string;
  white: string;
  transparentWhite_30: string;
  transparentWhite_60: string;
  lightBlueGradient: string;
  // TYPOGRAPHY
  p1Regular: string;
  p1Bold: string;
  p2Bold: string;
  h1: string;
  h3: string;
  h4: string;
}

export interface IFontStyle {
  fontSize: string;
  fontWeight: number | string;
  lineHeight: string;
}